pyproject.toml
[project]
name = "imageproc"
version = "0.1.0"
description = ""
authors = [
    {name = "eresh-mittal",email = "ereshmittal5@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "fastapi (>=0.115.10,<0.116.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "redis (>=5.2.1,<6.0.0)",
    "numpy (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
